--- a/src/rng.js
+++ b/src/rng.js
@@ -1,5 +1,7 @@
 import crypto from 'crypto';
 
+const rnds8 = new Uint8Array(16);
+
 export default function rng() {
-  return crypto.randomBytes(16);
+  return crypto.randomFillSync(rnds8);
 }