Performance is increased by using a single `rnds8` buffer to generate random numbers (as in browser version).

Changes affect only the `v1` and `v4`.

Benchmark master:

```
uuidv1() x 1,640,551 ops/sec ±0.76% (91 runs sampled)
uuidv1() fill existing array x 7,369,911 ops/sec ±0.49% (96 runs sampled)
uuidv4() x 389,271 ops/sec ±1.39% (92 runs sampled)
uuidv4() fill existing array x 407,681 ops/sec ±1.62% (91 runs sampled)
uuidv3() x 142,615 ops/sec ±0.81% (87 runs sampled)
uuidv5() x 141,386 ops/sec ±1.09% (87 runs sampled)
```

Benchmark this branch:

```
uuidv1() x 1,653,982 ops/sec ±0.91% (90 runs sampled)
uuidv1() fill existing array x 7,383,926 ops/sec ±0.66% (93 runs sampled)
uuidv4() x 471,989 ops/sec ±0.53% (93 runs sampled)
uuidv4() fill existing array x 484,766 ops/sec ±0.48% (96 runs sampled)
uuidv3() x 141,146 ops/sec ±1.18% (85 runs sampled)
uuidv5() x 147,054 ops/sec ±1.13% (90 runs sampled)
```